I have been neglecting git for some reason, most likely cause I don't care for this project. However, I recently added a ton of graphing power but recently the years fell off the Timeline which i will fix now.

+++ db/migrate/20140917004600_create_cars.rb
@@ -7,7 +7,7 @@ class CreateCars < ActiveRecord::Migration
       t.integer :city_mpg
       t.integer :highway_mpg
       t.integer :annual_fuel_cost
-      t.integer :avg_gas_price
+      t.float :avg_gas_price
 
       t.timestamps
     end
